Separate offchain payment datatypes from the general command datatypes
